refactor: component util fix
