+++ src/components/Ellipsis/Ellipsis.vue
@@ -1,6 +1,6 @@
 <script>
 import Tooltip from 'ant-design-vue/es/tooltip'
-import { cutStrByFullLength, getStrFullLength } from '@/components/_util/StringUtil'
+import { cutStrByFullLength, getStrFullLength } from '@/components/_util/util'
 /*
     const isSupportLineClamp = document.body.style.webkitLineClamp !== undefined;
